Add ServerPage and CreateServerPage components
